use nodejs pipeline construct for auto backoff in pipeline
changes the setup of the processing pipeline and limits the amount of concurrent queries to the database
we currently only store state at the end of a succesful pipeline and ignore members based on their timestamp while processing them.
