+++ app.ts
@@ -12,16 +12,18 @@ import {
   REPLACE_VERSIONS,
   RUNONCE
 } from "./config";
-import Consumer, { ConfigurableLDESOptions } from "./consumer";
+import { ConfigurableLDESOptions } from "./consumer";
+import LdesPipeline from "./ldes-pipeline";
 import { NamedNode } from "n3";
 let taskIsRunning = false;
 
 const consumerJob = new CronJob(CRON_PATTERN, async () => {
+  if (taskIsRunning) {
+    console.log("Another task is still running");
+    return;
+  }
   try {
-    if (taskIsRunning) {
-      console.log("Another task is still running");
-      return;
-    }
+    taskIsRunning = true;
     const endpoint = LDES_ENDPOINT_VIEW;
     if (endpoint) {
       const ldesOptions: ConfigurableLDESOptions = {
@@ -32,10 +34,14 @@ const consumerJob = new CronJob(CRON_PATTERN, async () => {
         ldesOptions.requestsPerMinute = LDES_REQUESTS_PER_MINUTE;
       }
       const datasetIri = new NamedNode(LDES_STREAM);
-      const consumer = new Consumer({ datasetIri, endpoint, ldesOptions });
+      const consumer = new LdesPipeline({ datasetIri, endpoint, ldesOptions });
       console.log("Started processing " + endpoint);
       await consumer.consumeStream();
       console.log("Finished processing " + endpoint);
+      if (RUNONCE) {
+        console.log("Job is complete.");
+        process.exit();
+      }
     } else {
       throw new Error("No endpoint provided");
     }
